chore: Move ownership controller into tracker package
